Changed SAM with FaceNet detector in Hue preprocessing pipeline
And also added a test for the preprocessing pipeline
